completed tab function
